Only kill process and drop db table if local
If RUNNING_LOCALLY env var not set, defaults to true. Allows distinction  between running tests locally and in CI pipeline.

+++ run_functional_tests.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+running_locally=${RUNNING_LOCALLY:-true}
+
 # Bring up the db and mock-sso docker containers
 docker-compose -f docker-compose.yaml -f docker-compose.override.yaml up -d
 
@@ -14,8 +16,9 @@ python defend_data_capture/manage.py testserver \
     & echo $! > backend.pid \
     & (sleep 5 && npx cypress run --headless --browser chrome --config-file false)
 
-# Kill the application using the pid previously saved
-kill $(cat backend.pid)
-
-# Drop the test database
-docker-compose exec db psql -h localhost -U postgres -c "DROP DATABASE IF EXISTS test_defend"
+if [ "$running_locally" == true ]; then
+    # Kill the application using the pid previously saved
+    kill $(cat backend.pid)
+    # Drop the test database
+    docker-compose exec db psql -h localhost -U postgres -c "DROP DATABASE IF EXISTS test_defend"
+fi
